framnet require optimizations/fixes
