+++ frameit/lib/frameit/config_parser.rb
@@ -1,3 +1,5 @@
+require_relative 'module'
+
 module Frameit
   class ConfigParser
     attr_reader :data
