+++ frameit/lib/frameit/runner.rb
@@ -1,6 +1,11 @@
 require 'deliver'
 require 'fastimage'
 
+require_relative 'frame_downloader'
+require_relative 'module'
+require_relative 'screenshot'
+require_relative 'device_types'
+
 module Frameit
   class Runner
     def initialize
